+++ src/utils.js
@@ -1,6 +1,6 @@
 import checker from "license-checker"
-import { basename } from "node:path"
-import { readFileSync } from "node:fs"
+import { basename, join } from "node:path"
+import { readdir, readFile } from "node:fs/promises"
 
 export function check ( start ) {
 	return new Promise(( resolve, reject ) => checker.init({ start }, ( err, pkgs ) => err ? reject(err) : resolve(pkgs)))
@@ -8,11 +8,11 @@ export function check ( start ) {
 
 export async function read ([ name, { licenses, licenseFile, path }]) {
 	const main = path == process.cwd()
-	const base = basename(licenseFile)
+	const base = licenseFile ? basename(licenseFile) : ""
 
 	if (base.toLowerCase() == "license" || base.endsWith(".txt") || /^license\-[^\.]+$/i.test(base)) {
-		const text = readFileSync(licenseFile).toString()
-		const license = toQuotes(text)
+		const text = await readFile(licenseFile)
+		const license = toQuotes(text.toString())
 
 		return {
 			name,
@@ -20,8 +20,8 @@ export async function read ([ name, { licenses, licenseFile, path }]) {
 			licenses,
 			license
 		}
-	} else if (base.toLowerCase() == "license.md" || base.toLowerCase().endsWith(".md")) {
-		const text = getFromMarkdown(licenseFile)
+	} else if (base.toLowerCase() == "license.md" || (base.toLowerCase().endsWith(".md") && base.toLowerCase() != "readme.md")) {
+		const text = await getFromMarkdown(licenseFile)
 		const license = !/\#/g.test(text) ? toQuotes(text) : text
 
 		return {
@@ -31,7 +31,8 @@ export async function read ([ name, { licenses, licenseFile, path }]) {
 			license
 		}
 	} else {
-		const license = await getFromReadme({ licenseFile, path })
+		const text = await getFromReadme({ licenseFile, path })
+		const license = text && !/\#/g.test(text) ? toQuotes(text) : text
 
 		return {
 			name,
@@ -46,21 +47,32 @@ function toQuotes ( text ) {
 	return text.replace(/^(\r?\n)+|(\r?\n)+$/, "").split(/r?\n/).map(( line ) => `> ${line}`).join("\n")
 }
 
-function getFromMarkdown ( path ) {
-	const text = readFileSync(path).toString()
-	const license = text.replace(/^ *\#/gm, "####")
+async function getFromMarkdown ( path ) {
+	const text = await readFile(path)
+	const license = text.toString().replace(/^ *\#/gm, "####")
 
 	return license
 }
 
-async function getFromReadme ({ licenseFile, path }) {
-	const text = readFileSync(licenseFile).toString()
-	const [, license = null ] = /\# license.*?\n([^]+)(\#|$)/i.exec(text) || []
+async function getFromReadme ({ path }) {
+	const readme = await getReadmePath(path)
 
-	if (license) {
-		return license
+	if (readme) {
+		const text = await readFile(join(path, readme))
+		const [, license = null ] = /\# (?:license|licensing).*?(?:\r?\n)+([^]+?)(\r?\n\#|$)/i.exec(text.toString()) || []
+
+		if (license) {
+			return license
+		} else {
+			// do stuff
+			return ""
+		}
 	} else {
-		// do stuff
 		return ""
 	}
+}
+
+async function getReadmePath ( dir ) {
+	const files = await readdir(dir)
+	return files.find(( name ) => name.toLowerCase() == "readme.md")
 }